new import
I installed the latest version of Stencyl to see if they fixed the bug with compiling and generating the game. The HTML5 version which is included in the free version of Stencyl is working. I was only able to import the older code into the newer version. I need to walk through every part of the game and fix lots of it. But I can at least compile it again.  swf files are too large to backup.